chore: split builders in different files
