+++ builders/index.js
@@ -1,9 +1,19 @@
 'use strict';
 /**
- * A DynamoDB client wrapper that simplifies writing requests and fetching responses.
+ * A DynamoDB expression builder that simplifies update items requests.
  * Let your AWS DynamoDB client take off ✈!
  *
  * @see https://flybondi.github.io/flynamo/
  */
 
-module.exports = require('./update-expression');
+const updateExpression = require('./update-expression');
+const put = require('./update-expression');
+const append = require('./update-expression');
+const remove = require('./update-expression');
+
+module.exports = {
+  updateExpression,
+  put,
+  append,
+  remove
+};
